+++ src/EditorFeatures/Core/EditorFeatures.csproj
@@ -110,7 +110,6 @@
     <Compile Include="FindUsages\SimpleFindUsagesContext.cs" />
     <Compile Include="Implementation\InlineRename\Dashboard\DashboardAutomationPeer.cs" />
     <Compile Include="Implementation\Structure\BlockTagState.cs" />
-    <Compile Include="Implementation\Suggestions\SuggestedActionsSource.cs" />
     <Compile Include="Tags\ExportImageMonikerServiceAttribute.cs" />
     <Compile Include="Implementation\NavigateTo\AbstractNavigateToItemDisplay.cs" />
     <Compile Include="CommandArgs.cs" />
